+++ lib/openapi_first/request_validation/request_body_validator.rb
@@ -7,21 +7,17 @@ module OpenapiFirst
         @operation = operation
       end
 
-      def validate!(runtime_request)
+      def validate!(parsed_request_body, request_content_type)
         request_body = operation.request_body
         return unless request_body
 
-        request_content_type = runtime_request.content_type
         schema = request_body.schema_for(request_content_type)
         RequestValidation.fail!(:header, status: 415) unless schema
 
-        parsed_request_body = runtime_request.body
         RequestValidation.fail!(:body) if request_body.required? && parsed_request_body.nil?
 
         validate_body!(parsed_request_body, schema)
         parsed_request_body
-      rescue BodyParser::ParsingError => e
-        RequestValidation.fail!(:body, message: e.message)
       end
 
       private